[Resolves #27] Improve api in class init (#29)
* Improve class __init__ functions

* require keyword-only arguments
* improve docstrings to clarify usage
* improve functionality of the update_docval utility function

* Backout new features of update_docval utility function

The original motivation was to remove docval items from class init functions
that we don't want to be part of the main user API because they are not
directly used by the user. However, it doesn't seem like this is possible
because they are being used when building a container from disk. So we're
backing-out the features to update_docval because they won't be used.

* Fix example in README

* Update release notes and add new contributor for most recently accepted PR

* Update changelog to include PR number following the example of pynwb

* Implement suggestions from PR review

* Fix docstring example